Refactor: modal state 훅 분리 작업
