+++ app/(route)/activity/edit/[id]/page.tsx
@@ -1,4 +1,6 @@
 import type { SubImage } from '@/_apis/activities/getActivity';
+import getActivity from '@/_apis/activities/getActivity';
+import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 
 import CommonLayout from '@/_components/CommonLayout';
 import StickyLayout from '@/_components/SideStickyLayout';
@@ -34,15 +36,22 @@ export interface ActivityEdit {
 /**
  * 체험 수정 페이지
  */
-function ActivityEdit({ params }: { params: { id: string } }) {
+export default async function ActivityEdit({ params }: { params: { id: string } }) {
   const { id } = params;
+  const queryClient = new QueryClient();
+
+  await queryClient.prefetchQuery({
+    queryKey: ['activity', id],
+    queryFn: () => getActivity(id),
+  });
+
   return (
     <CommonLayout>
       <StickyLayout>
-        <EditLayout id={id} />
+        <HydrationBoundary state={dehydrate(queryClient)}>
+          <EditLayout id={id} />
+        </HydrationBoundary>
       </StickyLayout>
     </CommonLayout>
   );
 }
-
-export default ActivityEdit;
