+++ app/(route)/activity/register/_components/BannerImage/index.tsx
@@ -6,6 +6,7 @@ import { useRouter } from 'next/navigation';
 import type { ActivityEdit } from '@/(route)/activity/edit/[id]/page';
 
 import { usePostImage } from '@/_hooks/use-post-image';
+import useModalState from '@/_hooks/useModalState';
 
 import type { Activity } from '../../page';
 import CommonModal from '../CommonModal';
@@ -18,25 +19,14 @@ export interface BannerImageProps<T> {
 
 /**
  * 메인 배너 이미지 업로드 및 form 데이터 업데이트하는 컴포넌트입니다.
+ *
  * @param value 체험 수정 페이지에서 이미 default 이미지값이 있는 경우 (이미지 주소)
  * @setFormdata 업로드한 이미지로 formdata 업데이트 (setState 함수)
  */
 function BannerImage<T extends Activity | ActivityEdit>({ value, setFormData }: BannerImageProps<T>) {
   const router = useRouter();
   const [bannerImage, setBannerImage] = useState<string[]>([]);
-  const [modalState, setModalState] = useState({
-    isOpen: false,
-    message: '',
-    onClose: () => {},
-  });
-
-  const closeModal = () => {
-    setModalState((prev) => ({
-      ...prev,
-      isOpen: false,
-    }));
-    modalState.onClose();
-  };
+  const { modalState, setModalState, closeModal } = useModalState();
 
   const getResponse = (res: string) => {
     setBannerImage([res]);
